+++ log/logger_set.go
@@ -9,7 +9,9 @@ import (
 // A set of log sinks (e.g., stdout, file, Sentry, etc.)
 // Logs are sent to all sinks
 type LoggerSet struct {
-	loggers []Logger
+	loggers             []Logger
+	minimumLevel        LogLevel
+	userPropertiesToLog *[]UserProperty
 }
 
 // Create a new log set, with the standard logger
@@ -28,16 +30,58 @@ func (l *LoggerSet) AddLogger(logger Logger) {
 
 // Convenience function to set the minimum log level for all
 // current log sinks.
-// Log sinks added after this call will not be affected
+//
+// Note: Log sinks added after this call will not be affected
 func (l *LoggerSet) SetMinimumLevel(logLevel LogLevel) {
 	for _, logger := range l.loggers {
 		logger.SetMinimumLevel(logLevel)
 	}
+	l.minimumLevel = logLevel
 }
+
+// Convenience function to get the minimum log level
+// for all current log sinks
+//
+// Note: Log sinks added after the most recent call of SetMinimumLevel
+// can have different minimum log levels.
+func (l *LoggerSet) GetMinimumLevel() LogLevel {
+	return l.minimumLevel
+}
+
+// Convenience function to set the user properties to log for all
+// current log sinks.
+//
+// Note: Log sinks added after this call will not be affected
+func (l *LoggerSet) SetUserPropertiesToLog(userPropertiesToLog *[]UserProperty) {
+	l.userPropertiesToLog = userPropertiesToLog
+
+	for _, logger := range l.loggers {
+		logger.SetUserPropertiesToLog(userPropertiesToLog)
+	}
+}
+
+// Convenience function to get the user properties to log
+// for all current log sinks
+//
+// Note: Log sinks added after the most recent call of SetUserPropertiesToLog
+// can have different values.
+func (l *LoggerSet) GetUserPropertiesToLog() *[]UserProperty { return l.userPropertiesToLog }
 
 func (l *LoggerSet) Log(level LogLevel, message string, err error, ctx context.Context) {
 	for _, logger := range l.loggers {
 		logger.Log(level, message, err, ctx)
+	}
+}
+
+func (l *LoggerSet) Logf(level LogLevel, err error, ctx context.Context, format string, args ...interface{}) {
+	for _, logger := range l.loggers {
+		logger.Logf(level, err, ctx, format, args...)
+	}
+}
+
+func (l *LoggerSet) Logln(level LogLevel, err error, ctx context.Context, args ...interface{}) {
+	for _, logger := range l.loggers {
+		logger.Logln(level, err, ctx, args...)
 	}
 }